# task-manager/backend-rest-api/src/swagger.yaml
openapi: 3.0.0
info:
  title: Task Manager API (₢ Daniel Souza)
  version: 1.0.0
  description: Each user has a profile (password = 123456) and can create tasks, which can be updated and deleted. The API is secured with JWT. The API is built with Node.js, Express and a fake database.
  author:
    name: Daniel Souza
components:
  securitySchemes:
    bearerAuth: # Define the security scheme for Bearer token
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /:
    get:
      summary: Get a greeting message
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              example:
                error: false
                message: Hello World!
  /signup:
    post:
      summary: User signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful signup
  /login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
  /users:
    get:
      summary: Get all users' profiles
      responses:
        '200':
          description: Successful response
  /profile:
    get:
      summary: Get user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
    put:
      summary: Update user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful update
    delete:
      summary: Delete user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful deletion
  /profile/tasks:
    get:
      summary: Get all user's tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
    post:
      summary: Create a new task for the user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Successful task creation
  /profile/tasks/{taskId}:
    get:
      summary: Get a user's task by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
    put:
      summary: Update a user's task by ID
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful task update
    delete:
      summary: Delete a user's task by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful task deletion